+++ app/controllers/ahoy_captain/application_controller.rb
@@ -75,6 +75,9 @@ module AhoyCaptain
     end
 
     def cached(*names)
+      if AhoyCaptain.cache.class == ActiveSupport::Cache::NullStore
+        return yield
+      end
       AhoyCaptain.cache.fetch("ahoy_captain:#{names.join(":")}:#{request.query_parameters.sort.map { |k,v| "#{k}-#{v}" }.join(":")}", expire_in: AhoyCaptain.config.cache.ttl) do
         yield
       end