+++ app/presenters/ahoy_captain/goals_presenter.rb
@@ -15,24 +15,27 @@ module AhoyCaptain
       selects = []
       last_goal = nil
       map = {}
+      total_event_count_by_name = @event_query.reselect("count(ahoy_events.name) as total").where(name: AhoyCaptain.config.goals.map(&:event_name)).group("ahoy_events.name").count("distinct ahoy_events.id")
       AhoyCaptain.config.goals.each do |goal|
-        queries[goal.id] = @event_query.select("count(distinct(visit_id)) as uniques, count(name) as total, name").where(name: goal.event_name).group("name")
-        selects << ["SELECT total, uniques, name from #{goal.id}"]
+        queries[goal.id] = @event_query.select("count(distinct(ahoy_events.visit_id)) as uniques, ahoy_events.name as name, count(distinct ahoy_events.id) as total").where(name: goal.event_name).group("ahoy_events.name")
+        selects << ["SELECT uniques, name, total from #{goal.id}"]
+
         map[goal.event_name] = goal
         last_goal = goal
       end
       select = selects.join(" UNION ").delete_suffix(" from #{last_goal.id}")
       select = select.delete_prefix("SELECT ")
       steps = ::Ahoy::Event.with(
-        queries
+        queries,
       ).select(select).from("#{last_goal.id}")
 
-      items = ::Ahoy::Event.with(steps: steps).select("total, uniques, name, 0 as conversion_rate").from("steps").index_by(&:name)
+      items = ::Ahoy::Event.with(steps: steps).select("total as total_events, 0 as total_visitors, uniques, name, 0 as conversion_rate").from("steps").index_by(&:name)
       @goals = []
       map.each do |name, _|
         if items[name]
           items[name].name = map[name].title
-          items[name].conversion_rate = ((items[name].total / total.to_d) * 100).round(2) * 100
+          items
+          items[name].conversion_rate = (((items[name].total_events / total_visitors.to_d) * 100) / 100.to_d).round(2)
           @goals << items[name]
         else
           @goals << OpenStruct.new(name: map[name].title, uniques: 0, total: 0, conversion_rate: 0)
@@ -42,8 +45,8 @@ module AhoyCaptain
       self
     end
 
-    def total
-      @total ||= @event_query.distinct(:visitor_token).count
+    def total_visitors
+      @total_visitors ||= @event_query.select(:visit_id).distinct.count
     end
 
     def as_json