+++ app/queries/ahoy_captain/stats/average_visit_duration_query.rb
@@ -3,7 +3,7 @@ module AhoyCaptain
     class AverageVisitDurationQuery < ApplicationQuery
       def build
         events = event_query
-                   .reselect("visit_id, max(time) - min(time) as visit_duration")
+                   .reselect("ahoy_events.visit_id, max(ahoy_events.time) - min(ahoy_events.time) as visit_duration")
                    .group("visit_id")
 
         ::Ahoy::Visit.joins("INNER JOIN #{::AhoyCaptain.event.table_name} ON #{::AhoyCaptain.event.table_name}.visit_id = visit_durations.visit_id")