+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
-## Version 0.69
+## Unreleased
 
-* Initial release
+### Added
+
+* Nothing, yet
+
+### Fixed
+
+* Fix visit and event query by properly merging the searches
+  - https://activerecord-hackery.github.io/ransack/going-further/merging-searches/
+  - this requires queries to specify the full table name 
+  
+### Changed
+
+* Changed the configuration file to properly reflect method names 
+
+## Version 0.77
+
+* Initial release and announcement thing